requirements.txt
python-dotenv
flask
flask-jwt-extended
flask-restx
flask-sqlalchemy
pyjwt<2.10
pandas
scikit-learn
numpy
matplotlib
jupyter
pandas
matplotlib
seaborn
flask_cors
mysql-connector-python-rf
pytest-html